pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lichess-tools"
version = "0.1.0"
description = "CLI tools for interacting with the Lichess API"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "httpx>=0.27",
    "rich>=13",
    "python-chess>=1.10",
    "tenacity>=8",
    "python-dateutil>=2.9",
    "tomli-w>=1.1",
    "tomllib; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
]

[project.scripts]
lichess = "lichess_tools.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/lichess_tools"]

[tool.ruff]
line-length = 100
target-version = "py312"
